move directories to ui
