--- app/src/main/java/com/traveloka/hotel/screens/search_hotel/SearchHotelScreen.kt
@@ -1,3 +1,0 @@
-package com.traveloka.hotel.screens.search_hotel
-
-class SearchHotelScreen
