+++ app/src/main/java/com/traveloka/hotel/ui/navigation/HotelScreens.kt
@@ -1,4 +1,4 @@
-package com.traveloka.hotel.navigation
+package com.traveloka.hotel.ui.navigation
 
 enum class HotelScreens {
     HomeScreen,
