+++ app/src/main/java/com/traveloka/hotel/ui/screens/register/RegisterScreen.kt
@@ -1,4 +1,4 @@
-package com.traveloka.hotel.screens.register
+package com.traveloka.hotel.ui.screens.register
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
