+++ app/src/main/java/com/traveloka/hotel/ui/screens/search_hotel/SearchHotelScreen.kt
@@ -1,0 +1,3 @@
+package com.traveloka.hotel.ui.screens.search_hotel
+
+class SearchHotelScreen
